// src/server/cms/types.ts
export interface PriceDistrictData {
	cityName: string;
	districtName: string;
	districtSlug: Slug;
	citySlug: string;
	cityType: "subdistricts" | "districts_onpage" | "districts_newpage" | "none";
	einwohner: number;
	state: Reference;
	stateName: string;
	stateSlug: string;
	kreis: string;
	isKreisfrei: boolean;
	seo: Seo;
	rating: Rating;
	sectionBuyingMarket: AddSection;
	sectionRentingMarket: AddSection;
	sectionHousePrices: AddSection;
	sectionApartmentPrices: AddSection;
	sectionHousePrices2: AddSection;
	sectionApartmentPrices2: AddSection;
	houseRent: RealEstatePrice[];
	houseBuy: RealEstatePrice[];
	apartmentRent: RealEstatePrice[];
	apartmentBuy: RealEstatePrice[];
	faqsList: Faq[];
	sources: BlockContent;
	neighboringCities: NeighboringCityPrices[];
	neighboringDistricts: NeighboringDistrictPrices[];
	subdistricts: NeighboringDistrictPrices[];
	sectionNeighboringCityBuyPrices: AddSection;
	sectionNeighboringCityRentPrices: AddSection;
	sectionBuyingMarket2: AddSection;
	sectionRentingMarket2: AddSection;
	sectionBuyingMarket3: AddSection;
	sectionRentingMarket3: AddSection;
	sectionDrivingFactors: AddSection;
	sectionIntro: AddSection;
	sectionPropertyPrices: AddSection;
}